fixed warning on solaris (incorect param type)
git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/trunk@1786 689a6050-402a-0410-94f2-e92a70836424

+++ strcommon.c
@@ -242,7 +242,7 @@ int escape_user(str *sin, str *sout)
 					(unsigned int)*p);
 			return -1;
 	    }
-	    if (isdigit(*p) || ((*p >= 'A') && (*p <= 'Z')) ||
+	    if (isdigit((int)*p) || ((*p >= 'A') && (*p <= 'Z')) ||
 				((*p >= 'a') && (*p <= 'z')))
 		{
 			*at = *p;
